requirements.txt
django==3.1.7
django-ckeditor==6.0.0
pillow==8.1.1
django-imagekit==4.0.2